Use lazy_loader for pyspark
